Se han añadido la instalación de canales y la ocupación que persiven los nodos, hace falta sacar graficas con estos nuevos ajustes. La siguiente mejora a tratar es la integración del algoritmo Bio-Inspirado tomando en cuenta el promedio ponderado de las variables como retardo, disponibilidad espectral y el uso de la bateria.

+++ Sim/My-tag.h
@@ -29,13 +29,13 @@ public:
   uint32_t GetNodeSinkId (); /*> Id del nodo que recibira el paquete*/
   u_long GetSEQNumber ();
   u_int32_t GetTypeOfPacket ();
-  uint8_t GetChanels();
+  uint64_t GetChanels();
   //void SetPosition (Vector pos);
   void SetNodeId (uint32_t node_id); /*> Id del nodo que envia*/
   void SetTimestamp (Time t);
   void CopySEQNumber (u_long number);
   void SetTypeOfpacket (int32_t TType);
-  void SetChanels(uint8_t chanels);
+  void SetChanels(uint64_t chanels);
   CustomDataTag ();
   virtual ~CustomDataTag ();
 
@@ -44,7 +44,7 @@ private:
   Time m_timestamp; /**> Tiempo en el que el paquete se creó*/
   u_long m_SEQNumber; /**> Numero de secuencia que identifica al paquete*/
   uint32_t m_TypeOfPacket; /**> 0 Paquete nuevo, 1 reenvio de paquete, 2 respuesta del Sink*/
-  uint8_t m_chanels;
+  uint64_t m_chanels;
 };
 } // namespace ns3
